updating polar fitting
