+++ py4DSTEM/process/polar/__init__.py
@@ -1,2 +1,3 @@
-from py4DSTEM.process.polar.polar_datacube import PolarDatacube, fit_amorphous_ring
+from py4DSTEM.process.polar.polar_datacube import PolarDatacube
+from py4DSTEM.process.polar.polar_fits import fit_amorphous_ring
